+++ tests/types/JsonSerializationTest.php
@@ -43,7 +43,7 @@ class JsonSerializationTest extends TestCase
     {
         $this::assertEquals(
             $object,
-            $object::createFromArray(\json_decode(\json_encode($object), true))
+            $object::createFromArray(\json_decode(\json_encode($object, \JSON_THROW_ON_ERROR), true))
         );
     }